Add GitLab CI pipeline status feedback
Show GitLab CI pipeline status for commits on GitLab.com so that I
don't have to check all the time.

Use `gl-util` (alpha) which I wrote for this purpose and is likely to
change to get the latest pipeline status. Use `async.zsh` to prevent
locking the prompt waiting for API response.